Refactor the EndpointSelectionStrategy code
